Incorrect throughput statement?
In the doc we mention "Throughput provisioned on an Azure Cosmos container is uniformly distributed across all the logical partitions of the container". It is my understanding that we provision throughput on the physical partition layer, not the logical (backed also by this doc: https://docs.microsoft.com/en-us/azure/cosmos-db/partition-data#physical-partitions). I believe this doc has an error unless I am missing something.